added settings to customize prescription default val

+++ src/Controller/AppointementController.php
@@ -10,6 +10,7 @@ use App\Entity\Appointement;
 use App\Entity\Patient;
 use App\Form\AppointementType;
 use App\Repository\AppointementRepository;
+use App\Repository\SettingsRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 use DateTimeImmutable as Date;
@@ -19,14 +20,35 @@ final class AppointementController extends AbstractController
   #[Route('/appointements', name: 'app_appointements')]
   public function index(AppointementRepository $ar, Request $request): Response
   {
-    $appointements = $ar->search($request->query->get('search'), $request->query->get('orderby'));
+    $u = $this->getUser();
+    if (!$u) {
+      return $this->redirectToRoute("app_login");
+    }
+
+    $appointements = $ar->search(
+      $request->query->get('search'),
+      $request->query->get('orderby'),
+      $request->query->get('from'),
+      $request->query->get('to')
+    );
     // dd($appointements);
     return $this->render('appointement/index.html.twig', compact('appointements'));
   }
 
   #[Route('/appointements/{id}/add', name: 'appointement_add')]
-  public function create(Request $request, Patient $patient, EntityManagerInterface $em)
+  public function create(Request $request, Patient $patient, EntityManagerInterface $em, AppointementRepository $ar, SettingsRepository $sr)
   {
+    $u = $this->getUser();
+    if (!$u) {
+      return $this->redirectToRoute("app_login");
+    }
+    $limit = $sr->find(1)->getLimitAppointements();
+    $appointementsCount = $ar->countTodaysAppointements();
+    if($appointementsCount === $limit){
+      $this->addFlash('error',"Appointements limit($limit) for today reached");
+      return $this->redirectToRoute("app_patients");
+    }
+
     $appointement = new Appointement();
     $form = $this->createForm(AppointementType::class, $appointement, [
       'is_edit' => false
@@ -55,6 +77,10 @@ final class AppointementController extends AbstractController
   #[Route('/appointements/{id}/edit', name: 'appointement_edit')]
   public function edit(Request $request, Appointement $appointement, EntityManagerInterface $em)
   {
+    $u = $this->getUser();
+    if (!$u) {
+      return $this->redirectToRoute("app_login");
+    }
     $form = $this->createForm(AppointementType::class, $appointement, [
       'is_edit' => true
     ]);
@@ -82,7 +108,10 @@ final class AppointementController extends AbstractController
     Appointement $appointement,
     EntityManagerInterface $em
   ): Response {
-
+    $u = $this->getUser();
+    if (!$u) {
+      return $this->redirectToRoute("app_login");
+    }
     if ($this->isCsrfTokenValid('delete_appointement_' . $appointement->getId(), $request->request->get('_token'))) {
       $em->remove($appointement);
       $em->flush();
